server/requirements.txt
# FastAPI app requirements
fastapi==0.88.0
uvicorn==0.20.0
python-dotenv==1.0.0
httpx==0.23.3
pydantic==1.10.12

# Flask app requirements (if using Flask instead)
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
llama-cpp-python==2.9.6

